Removed legacy google test related to the deleted interfaces
