+++ legacy_utest/stream_reduce_example_GT.cpp
@@ -22,82 +22,64 @@
 #include <fstream>
 #include <chrono>
 #include <stream_reduce.h>
-#include <string>
-#include <sstream>
 #include <gtest/gtest.h>
+
+using namespace std;
 using namespace grppi;
 
-std::vector<int> read_list(std::istream & is){
-  using namespace std;
-  std::vector<int> result;
-  string line;
-  is >> ws;
-  if(!getline(is,line)) return result;
-  istringstream iline{line};
-  int x;
-  while(iline >> x){
-    result.push_back(x);
-  }
-  return result;
-}
+int reduce_example3(auto& p){
 
-
-
-int stream_reduce_example(auto &p) {
-    using namespace std;
- 
 #ifndef NTHREADS
 #define NTHREADS 6
 #endif
 
-    ifstream is("txt/file.txt");
-    if (!is.good()) { cerr << "TXT file not found!" << endl; return 0; }
-
-    int reduce_var = 0;
-    
+    int total = 0;
+    int reduce_var=0;
+    std::vector<int> stream( 10000, 1 );
+    int index = 0;
+    int n=0;
     stream_reduce( p,
-        // GenFunc: stream consumer
-        [&]() {
-            auto r = read_list(is);
-            return ( r.size() == 0 ) ? optional<vector<int>>{} : optional<vector<int>>(r);
-        },  	
-        // TaskFunc: reduce kernel 
-        [&]( vector<int> v ) {
-            int loc_red = 0;
-            for( int i = 0; i < v.size(); i++ )
-                loc_red += v[i];
-            return loc_red;
+        // Reduce generator as lambda
+        [&]() { 
+            n++;
+            if(n != 1000000000) 
+              return (optional<int> ( 1 ));
+            else
+              return (optional<int> ());
         },
-        // RedFunc: final reduce
-        [&]( int loc_red ,int &reduce_var ) {
-            reduce_var += loc_red;
-    }, reduce_var
-);
-
-    return reduce_var;
+        //Window size
+        1000000,
+        1000000,
+        // Reduce kernel as lambda
+        std::plus<int>(),
+        // Reduce join as lambda
+        [&]( int a) {
+            total += a;
+        } 
+    );
+    return total;
 }
 
-
-TEST(GrPPI, stream_reduce_example_seq ){
+TEST(GrPPI, reduce_example3_seq ){
     sequential_execution p{};
-    EXPECT_EQ(5060408, stream_reduce_example(p) );
+    EXPECT_EQ(999999999, reduce_example3(p) );
 }
 
-TEST(GrPPI, stream_reduce_example_thr ){
+TEST(GrPPI, reduce_example3_thr ){
     parallel_execution_native p{NTHREADS};
-    EXPECT_EQ(5060408, stream_reduce_example(p) );
+    EXPECT_EQ(999999999, reduce_example3(p) );
 }
 
 #ifdef GRPPI_OMP
-    TEST(GrPPI, stream_reduce_example_omp ){
+    TEST(GrPPI, reduce_example3_omp ){
         parallel_execution_omp p{NTHREADS};
-        EXPECT_EQ(5060408, stream_reduce_example(p) );
+        EXPECT_EQ(999999999, reduce_example3(p) );
     }
 #endif
 #ifdef GRPPI_TBB
-    TEST(GrPPI, stream_reduce_example_tbb ){
+    TEST(GrPPI, reduce_example3_tbb ){
         parallel_execution_tbb p{NTHREADS};
-        EXPECT_EQ(5060408, stream_reduce_example(p) );
+        EXPECT_EQ(999999999, reduce_example3(p) );
     }
 #endif
 
@@ -107,3 +89,6 @@ int main(int argc, char **argv) {
     testing::InitGoogleTest(&argc, argv);
     return RUN_ALL_TESTS();
 }
+
+
+
